Fixes for Composite K-of-N implementation.
