+++ src/openssl/composite/composite_ameth.c
@@ -6,14 +6,14 @@
 #pragma GCC diagnostic ignored "-Wunused-function"
 
 // Local Include
-#include "composite_ameth.h"
+#include "composite_ameth_lcl.h"
 
 // ===============
 // Data Structures
 // ===============
 
 #ifndef _LIBPKI_COMPOSITE_OPENSSL_LOCAL_H
-#include "composite_ossl_internals.h"
+#include "composite_ossl_lcl.h"
 #endif
 
 // ======================
@@ -31,7 +31,7 @@
 // ==============================
 
 // Implemented
-int pub_decode(EVP_PKEY *pk, X509_PUBKEY *pub) {
+int pub_decode(EVP_PKEY *pk, X509_PUBKEY *pubkey) {
 
   // Strategy:
   //
@@ -60,13 +60,24 @@ int pub_decode(EVP_PKEY *pk, X509_PUBKEY *pub) {
   ASN1_BIT_STRING aBitStr;
     // Temp Octet Pointer
 
+  const void *params_value = NULL;
+    // Value for the parameters
+
+  const unsigned char *param_der = NULL;
+  int param_len = 0;
+    // Buffer
+
+  X509_ALGOR * pkey_alg = NULL;
+  int pkey_type = 0;
+    // Public Key Type and Algorithms
+
   // Input Checking
-  if (!pk || !pub) return 0;
+  if (!pk || !pubkey) return 0;
 
   // Let's use the aOctetStr to avoid the internal
   // p8 pointers to be modified
-  aBitStr.data = pub->public_key->data;
-  aBitStr.length = pub->public_key->length;
+  aBitStr.data = pubkey->public_key->data;
+  aBitStr.length = pubkey->public_key->length;
 
   // Gets the Sequence from the data itself, error if
   // it is not a sequence of ASN1_OCTET_STRING
@@ -81,18 +92,6 @@ int pub_decode(EVP_PKEY *pk, X509_PUBKEY *pub) {
   if ((comp_key = COMPOSITE_KEY_new()) == NULL) {
     PKI_ERROR(PKI_ERR_MEMORY_ALLOC, "Cannot allocate a new composite key internal structure");
     goto err;
-  }
-
-  // Checks the parameter, if present
-  if (pub->algor->parameter) {
-    if (pub->algor->parameter->type != V_ASN1_INTEGER) {
-      PKI_ERROR(PKI_ERR_PARAM_TYPE, "Expecting an INTEGER for the parameter");
-      return 0;
-    }
-    if ((comp_key->params = ASN1_INTEGER_dup(pub->algor->parameter->value.integer)) == NULL) {
-      PKI_ERROR(PKI_ERR_MEMORY_ALLOC, "Cannot duplicate the parameter INTEGER");
-      goto err;
-    }
   }
 
   // Process each component
@@ -143,6 +142,30 @@ int pub_decode(EVP_PKEY *pk, X509_PUBKEY *pub) {
   while ((aType = sk_ASN1_TYPE_pop(sk)) != NULL) {
     ASN1_TYPE_free(aType);
   } sk_ASN1_TYPE_free(sk);
+
+  // ======================
+  // Process Key Parameters
+  // ======================
+
+  // Retrieves the Public Key parameter
+  if (!X509_PUBKEY_get0_param(NULL, 
+                              &param_der, 
+                              &param_len, 
+                              &pkey_alg, 
+                              pubkey)) {
+    PKI_ERROR(PKI_ERR_X509_KEYPAIR_DECODE, "Cannot get the public key parameters");
+        return 0;
+  };
+
+  // Gets the type and the parameters
+  X509_ALGOR_get0(NULL, &pkey_type, &params_value, pkey_alg);
+  if (params_value) {
+    // If we have parameters, we need to save them to the key
+    if ((comp_key->params = ASN1_INTEGER_dup((ASN1_INTEGER *)params_value)) == NULL) {
+      PKI_ERROR(PKI_ERR_MEMORY_ALLOC, "Cannot duplicate the parameters");
+      goto err;
+    }
+  };
 
   // Assigns the key in the EVP_PKEY structure
   if (!EVP_PKEY_assign_COMPOSITE(pk, comp_key)) {
@@ -215,6 +238,10 @@ int pub_encode(X509_PUBKEY *pub, const EVP_PKEY *pk) {
   int buff_len = 0;
     // Temporary Storage for ASN1 data
 
+  ASN1_INTEGER * key_param = NULL;
+  int key_param_type = V_ASN1_UNDEF;
+    // K of N parameter
+
   // Input Checking
   if (!pub || !pk) return 0;
 
@@ -231,6 +258,7 @@ int pub_encode(X509_PUBKEY *pub, const EVP_PKEY *pk) {
   
   if ((sk = sk_ASN1_TYPE_new_null()) == NULL) {
     PKI_ERROR(PKI_ERR_MEMORY_ALLOC, "Cannot allocate a new stack of ASN1 Types");
+    if (key_param) ASN1_INTEGER_free(key_param);
     return 0;
   }
 
@@ -316,9 +344,23 @@ int pub_encode(X509_PUBKEY *pub, const EVP_PKEY *pk) {
   //   goto err;
   // }
 
+  // Encode the parameters (if any)
+  if (comp_key->params) {
+    fprintf(stderr, "[%s:%d] DEBUG: Detected k-of-n parameter, processing it...\n", __func__, __LINE__);
+    key_param_type = V_ASN1_INTEGER;
+    key_param = ASN1_INTEGER_dup(comp_key->params);
+    if (!key_param) {
+      PKI_ERROR(PKI_ERR_MEMORY_ALLOC, "Cannot duplicate the key parameters");
+      return 0;
+    }
+  }
+
+  fprintf(stderr, "[%s:%d] DEBUG: key_param_type = %d, key_param = %p\n", __func__, __LINE__, key_param_type, key_param);
+
   // We do not have parameters    
   if (!X509_PUBKEY_set0_param(pub, OBJ_nid2obj(pk->ameth->pkey_id),
-                              V_ASN1_INTEGER, ASN1_INTEGER_dup(comp_key->params), buff, buff_len)) {
+                        key_param_type, key_param, 
+                         buff, buff_len)) {
     PKI_ERROR(PKI_ERR_X509_KEYPAIR_ENCODE, "Cannot encode the parameter");
     goto err;
   }
@@ -329,6 +371,7 @@ int pub_encode(X509_PUBKEY *pub, const EVP_PKEY *pk) {
 err:
 
   // Free allocated memory
+  if (key_param) ASN1_INTEGER_free(key_param);
   if (buff && buff_len >= 0) OPENSSL_secure_clear_free(buff, (size_t) buff_len);
   if (bit_string) ASN1_BIT_STRING_free(bit_string);
   if (aType) ASN1_TYPE_free(aType);
@@ -498,6 +541,17 @@ int priv_decode(EVP_PKEY *pk, const PKCS8_PRIV_KEY_INFO *p8) {
   ASN1_OCTET_STRING outBitStr;
     // Temp BitString Pointer
 
+  const void *params_value = NULL;
+    // Value for the parameters
+
+  const unsigned char *param_der = NULL;
+  int param_len = 0;
+    // Buffer
+
+  const X509_ALGOR * pkey_alg;
+  int pkey_type = 0;
+    // Public Key Type and Algorithms
+
   // Input Checking
   if (!p8 || !pk) return 0;
 
@@ -555,6 +609,30 @@ int priv_decode(EVP_PKEY *pk, const PKCS8_PRIV_KEY_INFO *p8) {
       goto err;
     }
   }
+
+  // ======================
+  // Process Key Parameters
+  // ======================
+
+  // Retrieves the Public Key parameter
+  if (!PKCS8_pkey_get0(NULL, 
+                       &param_der, 
+                       &param_len, 
+                       &pkey_alg, 
+                       p8)) {
+    PKI_ERROR(PKI_ERR_X509_KEYPAIR_DECODE, "Cannot get the public key parameters");
+        return 0;
+  };
+
+  // Gets the type and the parameters
+  X509_ALGOR_get0(NULL, &pkey_type, &params_value, pkey_alg);
+  if (params_value) {
+    // If we have parameters, we need to save them to the key
+    if ((comp_key->params = ASN1_INTEGER_dup((ASN1_INTEGER *)params_value)) == NULL) {
+      PKI_ERROR(PKI_ERR_MEMORY_ALLOC, "Cannot duplicate the parameters");
+      goto err;
+    }
+  };
 
   // Free the stack memory
   while ((aType = sk_ASN1_TYPE_pop(sk)) != NULL) {
@@ -625,6 +703,10 @@ int priv_encode(PKCS8_PRIV_KEY_INFO *p8, const EVP_PKEY *pk) {
   unsigned char * buff = NULL;
   int buff_len = 0;
     // Temporary Storage for ASN1 data
+
+  ASN1_INTEGER * key_param = NULL;
+  int key_param_type = V_ASN1_UNDEF;
+    // K of N parameter
 
   // Input Checking
   if (!p8 || !pk) return 0;
@@ -715,9 +797,19 @@ int priv_encode(PKCS8_PRIV_KEY_INFO *p8, const EVP_PKEY *pk) {
   sk_ASN1_TYPE_free(sk);
   sk = NULL;
 
+  // Encode the parameters (if any)
+  if (comp_key->params) {
+    key_param_type = V_ASN1_INTEGER;
+    key_param = ASN1_INTEGER_dup(comp_key->params);
+    if (!key_param) {
+      PKI_ERROR(PKI_ERR_MEMORY_ALLOC, "Cannot duplicate the key parameters");
+      goto err;
+    }
+  }
+
   // Sets the params for the P8
   if (!PKCS8_pkey_set0(p8, OBJ_nid2obj(pk->ameth->pkey_id), 
-                       0, /* V_ASN1_UNDEF */ V_ASN1_NULL, NULL, buff, buff_len)) {
+                       0, key_param_type, key_param, buff, buff_len)) {
     PKI_ERROR(PKI_ERR_GENERAL, "Cannot set the P8 null parameters contents");
     goto err;
   }
@@ -728,6 +820,7 @@ int priv_encode(PKCS8_PRIV_KEY_INFO *p8, const EVP_PKEY *pk) {
 err:
 
   // Free allocated memory
+  if (key_param) ASN1_INTEGER_free(key_param);
   if (buff && buff_len >= 0) OPENSSL_secure_clear_free(buff, (size_t) buff_len);
   if (bit_string) ASN1_BIT_STRING_free(bit_string);
 
@@ -777,6 +870,9 @@ int priv_print(BIO *out, const EVP_PKEY *pkey, int indent, ASN1_PCTX *pctx) {
       BIO_printf(out, "        <NO TEXT FORMAT SUPPORT>\n");
     }
   }
+
+  BIO_printf(out, "Required Positive Components Validation (K-of-N): %d\n",
+    COMPOSITE_KEY_get_kofn(comp_key));
 
   return 1;
 }
@@ -1005,15 +1101,16 @@ int pkey_ctrl(EVP_PKEY *pkey, int op, long arg1, void *arg2) {
     case COMPOSITE_PKEY_CTRL_SET_K_OF_N: {
       // Sets the Valid Signature Requirement
       if (arg2) {
-        if (comp_key->params) ASN1_INTEGER_free(comp_key->params);
-        comp_key->params = ASN1_INTEGER_new();
-        ASN1_INTEGER_set(comp_key->params, *((int *)arg2));
+        if (PKI_OK != COMPOSITE_KEY_set_kofn(comp_key, *((int *)arg2))) {
+          PKI_ERROR(PKI_ERR_GENERAL, "Can not set the K of N value");
+          return 0;
+        }
       }
     } break;
 
     case COMPOSITE_PKEY_CTRL_GET_K_OF_N: {
       // Gets the Valid Signature Requirement
-      *(int *)arg2 = (int) ASN1_INTEGER_get(comp_key->params);
+      *(int *)arg2 = COMPOSITE_KEY_get_kofn(comp_key);
     } break;
 
     case ASN1_PKEY_CTRL_PKCS7_SIGN:
